Renamed sample files to match their region tags
