+++ google-cloud-storage/samples/storage_add_bucket_default_owner.rb
