+++ google-cloud-storage/samples/storage_print_bucket_acl.rb
