+++ google-cloud-storage/samples/storage_print_file_acl.rb
