+++ google-cloud-storage/samples/storage_remove_bucket_acl.rb
