+++ google-cloud-storage/samples/storage_remove_bucket_default_acl.rb
