+++ google-cloud-storage/samples/storage_remove_file_acl.rb
